+++ src/dataflow/src/decode/protobuf.rs
@@ -14,6 +14,7 @@ use repr::Row;
 #[derive(Debug)]
 pub struct ProtobufDecoderState {
     decoder: Decoder,
+    confluent_wire_format: bool,
     events_success: i64,
     events_error: i64,
 }
@@ -23,17 +24,38 @@ impl ProtobufDecoderState {
         ProtobufEncoding {
             descriptors,
             message_name,
+            confluent_wire_format,
         }: ProtobufEncoding,
     ) -> Self {
         let descriptors = DecodedDescriptors::from_bytes(&descriptors, message_name)
             .expect("descriptors provided to protobuf source are pre-validated");
         ProtobufDecoderState {
             decoder: Decoder::new(descriptors),
+            confluent_wire_format,
             events_success: 0,
             events_error: 0,
         }
     }
-    pub fn get_value(&mut self, bytes: &[u8]) -> Option<Result<Row, DecodeError>> {
+    pub fn get_value(&mut self, mut bytes: &[u8]) -> Option<Result<Row, DecodeError>> {
+        if self.confluent_wire_format {
+            // The first byte is a magic byte (0) that indicates the Confluent
+            // serialization format version, and the next four bytes are a big
+            // endian 32-bit schema ID.
+            //
+            // https://docs.confluent.io/current/schema-registry/docs/serializer-formatter.html#wire-format
+            if bytes.len() < 5 {
+                return Some(Err(DecodeError::Text(format!(
+                    "Confluent-style Protobuf datum is too few bytes: expected at least 5 bytes, got {}",
+                    bytes.len()
+                ))));
+            }
+            // For Protobuf, we just ignore the schema ID in the Confluent
+            // header. Unlike Avro, there's no concept of "resolving" a
+            // Protobuf reader schema against a writer schema. Instead we just
+            // have to trust that whoever wrote the data did so with a schema
+            // that is compatible with ours.
+            bytes = &bytes[5..];
+        }
         match self.decoder.decode(bytes) {
             Ok(row) => {
                 if let Some(row) = row {